added execution tests
