+++ tests/fixtures/configs/var2.yaml
@@ -1,9 +1,11 @@
+# test var declaration as inputs
+
 apiversion: orca/'1.0'
 
 version: '0.1'
 
 job:
-  - task: hydraulics
+  - task: example
     python: |
         assert inp == 0.025, "invalid inp {0}".format(inp)
         assert inp1 == 'st', "invalid st {0}".format(inp1)
